Sync blksize_t

+++ src/unix/notbsd/kfreebsd/other/b64/x86_64.rs
@@ -3,7 +3,7 @@
 pub type c_char = i8;
 pub type wchar_t = i32;
 pub type nlink_t = u64;
-pub type blksize_t = i64;
+pub type blksize_t = u32;
 pub type greg_t = i64;
 
 s! {
